requirements.txt
selenium==3.6.0
pyyaml==3.12
